requirements.txt
numpy
pandas
opencv-python
scikit-learn
torch
torchvision
Pillow
streamlit
tqdm
